ALU Operations

+++ src/memory/utils.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "system_constants.h"
+#include "../../include/system_constants.h"
 
 extern void print_data(MemoryDataType data, bool hex=true);
 
-extern int mem_dtype_to_int(MemoryDataType data);
+extern int64_t mem_dtype_to_int(MemoryDataType data);
 
-extern MemoryDataType int_to_mem_dtype(int data);
+extern MemoryDataType int_to_mem_dtype(int64_t data);
